Trimmed some dead files. Quick hack preliminary d3d11 renderer. I'm not likely to touch this again for quite some time. Sys_Error works properly on android, with an error message shown. Should be less of a stab in the dark if you get errors. Seg faults are still instantly fatal. Not much I can reliably do about those - most of the juicy ones will likely occur within the gl drivers (even if its my fault) and java will still call in to it. Stereoscopic (quad-buffered) rendering is in the gl renderer. You'll likely need a quatro to use it despite it being a gl 1.0 feature. No idea about ati. See r_stereo_method for non-quad-buffered alternatives. Tweaked networking to not overflow so much. Needs testing against other qw clients. Fixed an issue with surface numbers > 32k on limit-breaking maps. Fixed a preparse issue resulting in QW clients dying with the ne_ruins map's progs. Support for the DP-variant of BSP2. The depricated RMQ variant is still supported. QTV proxy now uses ipv6 hybrid sockets where possible. Preliminary pext support.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@4105 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/cl_tent.c
@@ -1587,7 +1587,7 @@ void CL_ParseTEnt (void)
 			dir[1] = MSG_ReadCoord();
 			dir[2] = MSG_ReadCoord();
 
-			cnt = MSG_ReadShort();	//count
+			cnt = (unsigned short)MSG_ReadShort();	//count
 			colour = MSG_ReadByte ();	//colour
 
 			P_RunParticleWeather(pos, pos2, dir, cnt, colour, "rain");
@@ -1612,7 +1612,7 @@ void CL_ParseTEnt (void)
 			dir[1] = MSG_ReadCoord();
 			dir[2] = MSG_ReadCoord();
 
-			cnt = MSG_ReadShort();	//count
+			cnt = (unsigned short)MSG_ReadShort();	//count
 			colour = MSG_ReadByte ();	//colour
 
 			P_RunParticleWeather(pos, pos2, dir, cnt, colour, "snow");
